[Refactor] - Language Selection : Moves UI classes used for building language selection spinner to media_ui module
