+++ audio/src/main/java/edu/artic/audio/AudioDetailsFragment.kt
@@ -25,6 +25,8 @@ import edu.artic.base.utils.updateDetailTitle
 import edu.artic.db.models.AudioFileModel
 import edu.artic.image.listenerAnimateSharedTransaction
 import edu.artic.media.audio.AudioPlayerService
+import edu.artic.language.LanguageAdapter
+import edu.artic.language.LanguageSelectorViewBackground
 import edu.artic.media.refreshPlayBackState
 import edu.artic.viewmodel.BaseViewModelFragment
 import io.reactivex.android.schedulers.AndroidSchedulers
