+++ media_ui/src/main/java/edu/artic/language/LanguageAdapter.kt
@@ -1,12 +1,12 @@
-package edu.artic.audio
+package edu.artic.language
 
 import android.graphics.Color
 import android.view.View
 import edu.artic.adapter.AutoHolderRecyclerViewAdapter
 import edu.artic.adapter.BaseViewHolder
 import edu.artic.adapter.DropDownAdapter
-import edu.artic.db.models.AudioFileModel
 import edu.artic.localization.SpecifiesLanguage
+import edu.artic.media.ui.R
 import kotlinx.android.synthetic.main.view_language_box.view.*
 
 /**
@@ -15,9 +15,9 @@ import kotlinx.android.synthetic.main.view_language_box.view.*
  * This is also responsible for creating the view seen at the top
  * of the list (i.e. the 'currently selected' language).
  */
-class LanguageAdapter : AutoHolderRecyclerViewAdapter<AudioFileModel>(),
+class LanguageAdapter : AutoHolderRecyclerViewAdapter<SpecifiesLanguage>(),
         DropDownAdapter<SpecifiesLanguage, BaseViewHolder> {
-    override fun View.onBindView(item: AudioFileModel, position: Int) {
+    override fun View.onBindView(item: SpecifiesLanguage, position: Int) {
         text.text = item.userFriendlyLanguage(context)
         text.setTextColor(Color.WHITE)
     }
