+++ media_ui/src/main/res/drawable/ic_language_caret.xml
