+++ media_ui/src/main/res/layout/view_language_box.xml
